Move stuff out of ColumnStat object.
